Move TopNftCollectionsRepository to 'topnftcollections' package
